Move FieldsMap to a function, LoggedEntry.ContextMap()
